+++ src/main/java/com/ono/omg/controller/ProductController.java
@@ -2,7 +2,7 @@ package com.ono.omg.controller;
 
 import com.ono.omg.dto.common.ResponseDto;
 import com.ono.omg.dto.request.ProductReqDto;
-import com.ono.omg.dto.response.ProductResDto;
+import com.ono.omg.dto.response.ProductResponseDto;
 import com.ono.omg.repository.product.ProductRepository;
 import com.ono.omg.security.user.UserDetailsImpl;
 import com.ono.omg.service.ProductService;
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+import static com.ono.omg.dto.response.ProductResponseDto.*;
 import static com.ono.omg.dto.response.ProductResponseDto.MainPageResponseDto;
 
 @RestController
